Refactor cluster api tests: move common code to api_testing_common
`parse_command_line` function has been moved
